wait...

+++ database/migrations/2020_08_13_183050_create_photos_table.php
@@ -15,11 +15,11 @@ class CreatePhotosTable extends Migration
     {
         Schema::create('photos', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
             $table->string('uri');
             $table->boolean('public');
-            $table->integer('height');
-            $table->integer('width');
+            $table->bigInteger('height');
+            $table->bigInteger('width');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->timestamps();
         });
